// packages/design-system/src/utilities/_deprecated_backups/generics_v1_backup/types/component-type.type.ts
/**
 * Component type mapping for prefix generation
 */
export type ComponentType =
    | 'button'
    | 'btn'
    | 'input'
    | 'textarea'
    | 'card'
    | 'checkbox'
    | 'radio'
    | 'switch'
    | 'select'
    | 'field'
    | 'label'
    | 'typography'
    | 'text'
    // TODO: Add these when CSS files are created
    // | 'badge'
    // | 'alert'
    | 'modal'
    | 'drawer'
    | 'tooltip'
    | 'accordion'
    | 'tab'
    | 'progress'
    | 'avatar'
    | 'chip'
    | 'divider'